feat(exec_simplecmd.c): get_fullpath()をutils.cに移動
